// communityAdminBackend/libs/audit/enums/auditParam.ts
export enum AuditParam {
  REQUESTID = 'RID',
  USER_AGENT = 'UserAgent',
  REFERER = 'Referer',
  RESPONSE_STATUS = 'ResponseStatus',
  REQUEST = 'Request',
  HTTP_METHOD = 'HttpMethod',
  RESPONSE = 'Response',
  URL = 'Url',
  OUT_URL = 'OutUrl',
  ELAPSED = 'Elapsed',
  HOST = 'Host',
  SERVER_NAME = 'ServerName',
  MULTIPART_REQUEST = 'MultipartRequest',
  REQUESTNAME = 'RequestName',
  CONSUMER = 'Con',
  VERSION = 'Ver',
  X_HOST = 'X-Host'
}